APKs can now be downloaded automatically. Moved APK downloads to the background service, renamed WebService. Worked on cleaning up old downloads. Removed 9Folders update source because of anti-scraping protections.
