+++ build.gradle
@@ -40,7 +40,7 @@ android {
         minSdkVersion 15
         targetSdkVersion 23
         versionCode 18
-        versionName "2.0.5"
+        versionName "2.1.0"
     }
 
     // Automatic signing for release builds. Credentials and keystore are not present in the VCS.
