+++ src/main/java/fr/kwiatkowski/apktrack/service/BroadcastHandler.java
@@ -82,12 +82,8 @@ public class BroadcastHandler extends BroadcastReceiver
             }
             else // The app was "added", but already exists. This is actually a replacement.
             {
-                // If ApkTrack had downloaded an APK for the app, it is now useless. Delete it.
-                if (app.get_download_id() != 0) {
-                    app.clean_downloads(ctx);
-                }
-                else if (!InstalledApp.detect_new_version(ctx.getPackageManager(), package_name)) {
-                    return; // Do not send an event if the version number hasn't changed and no APK was deleted.
+                if (!InstalledApp.detect_new_version(ctx, package_name)) {
+                    return; // Do not send an event if the version number hasn't changed.
                 }
                 type = ModelModifiedMessage.event_type.APP_UPDATED;
             }
