+++ src/main/java/fr/kwiatkowski/apktrack/service/message/StickyUpdatedMessage.java
@@ -19,6 +19,6 @@ package fr.kwiatkowski.apktrack.service.message;
 
 /**
  * Empty object which is only used as an event type.
- * It is used to notify the WebScraperService that the notification should be updated.
+ * It is used to notify the WebService that the notification should be updated.
  */
 public class StickyUpdatedMessage {}
